main -> fpga_main, 2 space indent and define for buffer header size
