+++ interface/Makefile
@@ -1,0 +1,178 @@
+# CMAKE generated file: DO NOT EDIT!
+# Generated by "Unix Makefiles" Generator, CMake Version 3.9
+
+# Default target executed when no arguments are given to make.
+default_target: all
+
+.PHONY : default_target
+
+# Allow only one "make -f Makefile2" at a time, but pass parallelism.
+.NOTPARALLEL:
+
+
+#=============================================================================
+# Special targets provided by cmake.
+
+# Disable implicit rules so canonical targets will work.
+.SUFFIXES:
+
+
+# Remove some rules from gmake that .SUFFIXES does not remove.
+SUFFIXES =
+
+.SUFFIXES: .hpux_make_needs_suffix_list
+
+
+# Suppress display of executed commands.
+$(VERBOSE).SILENT:
+
+
+# A target that is always out of date.
+cmake_force:
+
+.PHONY : cmake_force
+
+#=============================================================================
+# Set environment variables for the build.
+
+# The shell in which to execute make rules.
+SHELL = /bin/sh
+
+# The CMake executable.
+CMAKE_COMMAND = /usr/bin/cmake
+
+# The command to remove a file.
+RM = /usr/bin/cmake -E remove -f
+
+# Escaping for special characters.
+EQUALS = =
+
+# The top-level source directory on which CMake was run.
+CMAKE_SOURCE_DIR = /home/adwe/Documents/WS17/Project/PineappleExpress/interface
+
+# The top-level build directory on which CMake was run.
+CMAKE_BINARY_DIR = /home/adwe/Documents/WS17/Project/PineappleExpress/interface
+
+#=============================================================================
+# Targets provided globally by CMake.
+
+# Special rule for the target rebuild_cache
+rebuild_cache:
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Running CMake to regenerate build system..."
+	/usr/bin/cmake -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR)
+.PHONY : rebuild_cache
+
+# Special rule for the target rebuild_cache
+rebuild_cache/fast: rebuild_cache
+
+.PHONY : rebuild_cache/fast
+
+# Special rule for the target edit_cache
+edit_cache:
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Running CMake cache editor..."
+	/usr/bin/ccmake -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR)
+.PHONY : edit_cache
+
+# Special rule for the target edit_cache
+edit_cache/fast: edit_cache
+
+.PHONY : edit_cache/fast
+
+# The main all target
+all: cmake_check_build_system
+	$(CMAKE_COMMAND) -E cmake_progress_start /home/adwe/Documents/WS17/Project/PineappleExpress/interface/CMakeFiles /home/adwe/Documents/WS17/Project/PineappleExpress/interface/CMakeFiles/progress.marks
+	$(MAKE) -f CMakeFiles/Makefile2 all
+	$(CMAKE_COMMAND) -E cmake_progress_start /home/adwe/Documents/WS17/Project/PineappleExpress/interface/CMakeFiles 0
+.PHONY : all
+
+# The main clean target
+clean:
+	$(MAKE) -f CMakeFiles/Makefile2 clean
+.PHONY : clean
+
+# The main clean target
+clean/fast: clean
+
+.PHONY : clean/fast
+
+# Prepare targets for installation.
+preinstall: all
+	$(MAKE) -f CMakeFiles/Makefile2 preinstall
+.PHONY : preinstall
+
+# Prepare targets for installation.
+preinstall/fast:
+	$(MAKE) -f CMakeFiles/Makefile2 preinstall
+.PHONY : preinstall/fast
+
+# clear depends
+depend:
+	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 1
+.PHONY : depend
+
+#=============================================================================
+# Target rules for targets named fpga_main
+
+# Build rule for target.
+fpga_main: cmake_check_build_system
+	$(MAKE) -f CMakeFiles/Makefile2 fpga_main
+.PHONY : fpga_main
+
+# fast build rule for target.
+fpga_main/fast:
+	$(MAKE) -f CMakeFiles/fpga_main.dir/build.make CMakeFiles/fpga_main.dir/build
+.PHONY : fpga_main/fast
+
+fpga_main.o: fpga_main.cpp.o
+
+.PHONY : fpga_main.o
+
+# target to build an object file
+fpga_main.cpp.o:
+	$(MAKE) -f CMakeFiles/fpga_main.dir/build.make CMakeFiles/fpga_main.dir/fpga_main.cpp.o
+.PHONY : fpga_main.cpp.o
+
+fpga_main.i: fpga_main.cpp.i
+
+.PHONY : fpga_main.i
+
+# target to preprocess a source file
+fpga_main.cpp.i:
+	$(MAKE) -f CMakeFiles/fpga_main.dir/build.make CMakeFiles/fpga_main.dir/fpga_main.cpp.i
+.PHONY : fpga_main.cpp.i
+
+fpga_main.s: fpga_main.cpp.s
+
+.PHONY : fpga_main.s
+
+# target to generate assembly for a file
+fpga_main.cpp.s:
+	$(MAKE) -f CMakeFiles/fpga_main.dir/build.make CMakeFiles/fpga_main.dir/fpga_main.cpp.s
+.PHONY : fpga_main.cpp.s
+
+# Help Target
+help:
+	@echo "The following are some of the valid targets for this Makefile:"
+	@echo "... all (the default if no target is provided)"
+	@echo "... clean"
+	@echo "... depend"
+	@echo "... rebuild_cache"
+	@echo "... fpga_main"
+	@echo "... edit_cache"
+	@echo "... fpga_main.o"
+	@echo "... fpga_main.i"
+	@echo "... fpga_main.s"
+.PHONY : help
+
+
+
+#=============================================================================
+# Special targets to cleanup operation of make.
+
+# Special rule to run CMake to check the build system integrity.
+# No rule that depends on this can have commands that come from listfiles
+# because they might be regenerated.
+cmake_check_build_system:
+	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 0
+.PHONY : cmake_check_build_system
+
